rename: GameData, Player를 domain 패키지로 변경
