+++ src/main/java/bridge/BridgeGame.java
@@ -1,5 +1,7 @@
 package bridge;
 
+import bridge.domain.GameData;
+
 /**
  * 다리 건너기 게임을 관리하는 클래스
  */
