+++ src/main/java/bridge/domain/CorrectResult.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.domain;
 
 public enum CorrectResult {
     CORRECT("O"),
