+++ src/main/java/bridge/domain/GameData.java
@@ -1,4 +1,6 @@
-package bridge;
+package bridge.domain;
+
+import bridge.PlayerCommend;
 
 import java.util.ArrayList;
 import java.util.List;
